Improve colonisation.
In case an outpost ship reaches a system, try to colonise the system
first. Only if the system cannot be colonised, consider further systems.

This fixes the issue of moving an outpost ship to another system, even
though the current system can be colonised, just because another outpost
ship now is targeting the current system.